+++ common/instructions/return.h
@@ -7,7 +7,7 @@ class Return : public Instruction
 {
 public:
     static const std::string MNEMONIC;
-    void operator()(std::vector<float>& memory, uint8_t& subprogram_index, std::vector<uint16_t>& instruction_addresses,
+    void operator()(std::vector<int32_t>& memory, uint8_t& subprogram_index, std::vector<uint16_t>& instruction_addresses,
         std::vector<uint8_t>& return_indices) const override;
     std::vector<std::string> toStringTokens() const override;
     static std::unique_ptr<Instruction> fromStringTokens(const std::vector<std::string>& tokens);
